backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.8.0
python-multipart>=0.0.6
pandas>=2.1.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
pymupdf>=1.23.0
openai>=1.12.0
duckduckgo-search>=4.1.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
rapidocr-onnxruntime>=1.4.0
opencv-python>=4.5.1,<4.11
python-docx>=1.1.0
python-pptx>=1.0.0
bcrypt>=4.0.0
readability-lxml>=0.8.1
akshare>=1.12.0
